Create LICENSE
Intention is that any future use or fork of this repo (including any future or past changes) will also be free and open to all. Any uses of this software MAY NOT charge any fee or generate any profit. All uses of this software must be FREE for all users under all circumstances.